i have fixed problem food analayz not working and also i have fixed workout generator and caloriy calculator

+++ lib/data/models/calorie_result.dart
@@ -1,6 +1,6 @@
 class CalorieResult {
-  final double Bmr;
-  final double Tdee;
+  final int Bmr;
+  final int Tdee;
   final int RecommendedCalories;
 
   final List<dynamic> note;
